Update grpc-go to 1.5.2, and minor changes to dial
Grpc docs suggest that using WithBlock in a dial call is not a good
pattern (https://github.com/grpc/grpc-go/blob/master/Documentation/anti-patterns.md).

This CL changes the dial function in ipc to remove that dial option.
However, bootstrap depends on WithBlock to verify that reproxy has
started correctly, so a second function is added to include WithBlock
that is used for that particular situation.

This is driven by an issue cropping up in cog (b/289391270) which seems
to be coming from grpc.  With the version update and the changes to
dialing, it might be mitigated (don't have enough data yet to confirm or
deny that), but we should probably be doing the connection this way
anyway.

Bug: 289391270
Test: Updated idletimeout
Change-Id: Ie52575fa2d4a6749fa3e22a269c1672b146f62c9
GitOrigin-RevId: 74f140d46e8396b99850bbe45946e6dcc455b726