Add ExperimentalExitOnStuckActions feature
The feature allows to close reproxy with exit_code=1
if execution of any of build actions exceeded 2*reclient_timeout.
The feature is controlled by exit_if_reclient_timeout_exceeded flag,
and will enable determining why certain actions get stuck for
for duration much longer than reclient_timeout.

Test:
Artificially changed the timeout duration in the code and verified
that without the flag set it doesn't get triggered, but with the flag
set it shuts down reproxy and produces the call stacks.

Bug: b/284246561
Change-Id: I7a683d641053597492a0bf4ef360af1784fac2f9
GitOrigin-RevId: 6c414d868555bc020f8cc18ce6d93dccf53439fc