Make reproxy downloads atomic.
This CL changes reproxy's behavior to ensure atomic writes of outputs.
This is done by writing outputs to a temporary location and moving them
to the final destination, which should be atomic. The temporary location
should be on the same physical device as the final output, and that
should be set using the new flag download_tmp_dir. If the flag is not
set, reproxy will download the temporary artifacts to a subdirectory of
the action's working directory. Typically that would be under the out
directory for a ninja build.

Bug: b/299198127
Test: unit tests, chrome build on specialist workstation. Build time
unaffected by this change (~3.5m for building `chrome`).

Change-Id: I00348db625d302578aaa0c39f2f4b2d2a7082090
GitOrigin-RevId: 57130a78d96df74489f76f01d25f6d67ac98499b