Stash files from input-output dirs in compare mode
Currently, when a directory is both an input and an output, we don't
stash the files inside that directory when rerunning the action. This
can potentially cause incorrect action outputs.

This change ensures that we iterate through files within input/output
directories and includes them to the list of files to stash prior to
rerunning the action locally.

Bug: b/307932183
Tested: Updated unit tests and tested with art/android builds where
compare mode was previously failing.

Change-Id: I9409df599da24fb96123a7bcc560d8b0f6c6249c
GitOrigin-RevId: 9e065c535d6984fca1a97f7f69f008553fa82569