Handle virtual inputs for various include-type paths containing '..'
Handles figuring out virtual inputs for a number of include-type flags,
and handles getting maximal subdirectories out of the path so remote
clang will be able to traverse the path properly.

NOTE: possible performance impact, this is increasing the processing on
every path that comes from on of these flags.

Bug: b/311460875
Test: Unit tests updated
Change-Id: I0585767f9aa3adb37dfc746c8ceedf3c26e4837a
GitOrigin-RevId: b172d6193132f7748d134b5236012720f7e85ecb