Ensure resource manager has at least 1 cpu to work with.
If you have a low-core machine and a small resource fraction set, you
can get cases where the resource manager thinks there are zero cpus.
But it caps requests with more cpu than is actually available to the
minimum of both.  And if it thinks one is zero, then any attempt to
acquire the semaphone does so without waiting because the weight is
zero.

Bug: b/309944793
Test: Unit tests updated
Change-Id: If7bb524fc9a458ad141ed665fe15a83c725745e6
GitOrigin-RevId: fbf493c8743f3fab4791e9e7fc1111734d92b077