Replace all log.Fatalf in rewrapper/bootstrap with log.Exitf
Fatalf() prints a stack trace that is never useful for the user to see.
We should avoid printing the stack trace and only print the error
message we would like to print.
log.Exitf() still prints the line number from which this is called,
which will still be useful to see (as opposed to just printing the
message to stderr and exiting with an error code.

TESTED=ran rewrapper locally and tested the error messages and exit code
is non-zero.
Bug: n/a

Change-Id: Ia3c110350f8cf39f5adec707821e55211509975b
GitOrigin-RevId: 89d5946ee441eab0851ccc9fedf3f608f384ef59