Add transition to allow scandeps to build with clang-cl in a mingw build
clang-cl was chosen instead of msvc because clanscandeps will also build
out of the box with clang-cl whereas there were issues with msvc
The transition on the platform flag will be removed in a later cl once
all cgo code is removed and go binaries no longer need mingw on windows

We need to transition on --cxxopts becuase -std=c++{14,17} is invalid on
clang-cl so we need to use /std:c++{14,17}, this causes non clang-cl
compilers to throw an error. This needs to be a transition and cannot be
specified as copts in the cc_binary target as we need this to be applied
to all dependencies.

Test: Presubmits pass and still build
Bug: b/314967604

Change-Id: I5f94abb6cd9f466f8fd11c5db0082b80299add1b
GitOrigin-RevId: 2cd4a2ecbee6fad4eaff688dd69a244fd0cb4ae4