Refactor scanner to preserve prev flag parse result
Refactor the scanner struct so that it keep the prev flag parse result
in memory.

Background: Currently, Clang-CL flagparser is the only flagparser that need to have access to the prev FlagResult, because in case `prev.Values[0] == "-MF" && fcurr.NormalizedKey == "-clang"`, it need to pop the last flag. To make Clang-CL support @rsp file, we need to use `handleArgFunc()` inside of `handleClangCLFlags()`. To expose the prev to handleArgFunc(), without defining a global prev variable in clangcl/flagparser.go (see patch 3 of this CL: http://tg/c/2156202/3), we decide to refactor scanner to make scanner to keep prev as a field.

This CL is the pre-requirement for the chained CL: http://tg/2156202

Test: Unit test
Bug: b/320984781
Change-Id: Ia40661881c662bdcfe4e455ac5f9809f42fa337a
GitOrigin-RevId: f3b0dade32766f2d5ff92b9233185bf2f234ac39