Fix bug in remoteRacing
Currently, if an action got error in execContext.Result.Err, we
treat it as a cache hit (because execContext.Result is not nill),
then, we might wait for max 60s for downloading the remote result before trying to run it locally.

This CL fix this bug, so that if the checkCache returns a not nil and not cache hit result (basically an error status), we will not wait for downloading anything from remote anymore, and directly start local execution.

Bug: b/326080269
Test: Unit test
Change-Id: I93c7287a226fe898e9960c0240c0864b9a7732a4
GitOrigin-RevId: 01fa5559c83bbd08e390c7d3048cbadae2c3deb1