Apply filters for downloads
We introduce a flag, `download_regex`, to select/exclude which output
files to be downloaded.

Depending on the first char of this flag's value, there are two types
regex:

1) a negative regex is a regex str start with "-", -<regex>, any
output file matches this regex will not be downloaded;

2) a positive regex is just a regex str not start with "-", <regex>, any
output file matches this regex will be downloaded.

To not breaking current interface, if the filter is "" or "-", or cannot
 be compiled as regex, such as "*" or "-*" we are going to fallback to
 --download_output flag.

pb.go files under api directory is updated with `./scripts/regenpbgo.sh`

Bug: b/327611049
Test: Unit test
Change-Id: I3a6edf718640f623b170edb829852c8b6f453167
GitOrigin-RevId: e1e4ea82c65f71cbf7169483a1e0978e1ae873f2