Adjust max listen size for messages from rewrapper/siso
Having the max message size for the proxy listener be 32mb turns out to
cause pretty significant performance problems on intel macs.

Reduce the max listen size to 8mb from 32mb, which should still be big
enough to address the initial cause for increasing the size in the
first place.

Bug: b/330893671
Test: presubmits
Change-Id: I601794f2870a2d163257b5c2f7f49ffac7c177ad
GitOrigin-RevId: 32d3b4e54bd7510fcb5e16d829b7c201f0419f8c