Remove default server address of 127.0.0.1:8000
A big cause of problems is due to trying to run a one off action outside
the normal build flow, like using ninja to try and re-run a failed
command.  Because of this default, rewrapper spins its wheels triyng to
connect to this address, even though it has nothing at all to do with
the build config.  Remove this as the default and make it empty.

Also actually add checks for this being empty in reproxy and bootstrap,
I'm surprised they were never there in the first place.

Bug: b/339176563
Test: presubmits
Change-Id: Ie2de32ab483342195df19461289fe2feeacd6a54
GitOrigin-RevId: 899432c7d4da54f2d64c5e867b6be025cad3bf05