Add support for RCLE mode (Remote Cache Local Execution).
In RCLE mode, if the action was previously executed, reproxy pulls the
cache hit, but if its a cache miss, reproxy will locally execute the
action and not attempt to cache it. This relies on the action key not
embedding the reproxy version in RCLE mode, as opposed to LERC mode
where cache writing can introduce correctness risk.

Bug: b/209059608
Test: ran action with remote mode, then local /
remote_update_cache=false, second action got a cache hit.

Change-Id: Ibe4e814cf69e5269d54d3e67c5f6c3d42078f043
GitOrigin-RevId: 48621454b725aeb72baa7f4da16af665d6a5981c