Evaluate symlinks in working directory.
If symlinks are present in the working directory path, rewrapper will
think that the path is not under the exec root and fail with an error
message like the following:

F0613 14:21:24.663747 1298125 main.go:164] Current working directory ("/noncanonical/path/out") is not under the exec root ("/canonical/path/"), relative working dir = "../../noncanonical/path/out"

Fix it by resolving symlinks in the path. Because this also handles
the existing /proc/self/cwd case, remove the code handling that case.

Change-Id: I81c9a4f8a1c87c7fa1f189d3e0a3118cacc064f6
Bug: 347963344
Test: verified that build fails with a permission error (due to no []
GitOrigin-RevId: cd91ea63496765f2d4f8f2f22696ec7f4cfd2bdc