Pipe peak num of bq uploader to reproxy.INFO
This CL pass the peak num of bigquery uploaders to reproxy.INFO log.

This metric will help diagnose bottlenecks in the BigQuery upload process.

In general, for every build, even on small 8 core machine, there are about 180 seconds at the end of the build that the peak num of actions is constantly almost zero. In the future, we can dynamically adjust the number of BigQuery uploaders based on the in-flight actions to better utilize this available machine capacity.

Test: run a local chrome build, and check the records are saved in bq.
Bug: b/345015849
Change-Id: If2a0e4ec5368dea3db91a81b122db2af2ae98639
GitOrigin-RevId: d43da69ec0f75694868f0bdaf7213456cbf3e7a9