Update github ci to use macos-14
Macos-14 runners are arm only, so now there is a native arm build and a
cross compiled x86 build.
Updating solved the c-ares build issue on gitub ci, so since cache
poisining was not the cause this also re-enables shared caches for mac builds.
Also enabled continue-on-error to allow all builds to complete even if
one fails.

Test: https://github.com/bazelbuild/reclient/actions/runs/9998906494?pr=70
Bug: N/A
Change-Id: Iae28fdb3b4b3fb3d6009d171d9a2f401972ede24
GitOrigin-RevId: 1e21a85336a0e0ecebd74170c0071b40b9843bc3