Use UDS for reproxy <-> depsscanner communciation on Windows
UDS is now supported in Windows and this change is intending to get rid
of TCP port communication since it seems mildly flaky on Windows CQ
machines for Chrome.

The UDS socket will be put inside the temp folder in Windows.

Bug: b/352206623
Tested: Ran the server on Windows and ensured we are able to connect to it
Change-Id: Ia9f80ec6ee7910ae53f1dccfce7bbc5b98568591
GitOrigin-RevId: 76543e3489c182fede7feed4c36abd66373e34ea