Fix bigquery translator for mismatches
local_digest and remote_digest are deprecated, and we currently only use local_digests and remote_digests.

We need to explicitly make the repeated string field to be an empty slice of string to upload to bigquery.

In Android CI, there is one builder fail to upload to bigquery due to this issue. See b/358569375#comment6

Test: Unit test
Bug: b/358569375
Change-Id: I874fc00c09b0114361b397903d3cf01ac88f7471
GitOrigin-RevId: 48ed3aecee55f62dd9e69916690437fb482a3e8c