Patch LLVM/Clang to support memtag-stack
Previously, we pypassed the clang flag `-fsanitize=memtag-stack` so that `m libc` was able to build with these two CLs (aosp/3261733, aosp/3264014).

But just found out today that this is not enough:

 1. `m libbase` was still falling back due to clang flag like this:
 `-fsanitize=integer,integer,memtag-stack`

 2. There are also some other targets depending on this flag (see
 example: https://paste.googleplex.com/4516018109284352)

This CL patched this github CL (http://shortn/_QQURhHpCQy) into reclient's input processor, so that re-client officially supports
this memtag-stack clang flag.

Test: abtd run: 1) caiman_fullmte_userdebug 8 fallbacks (http://shortn/_yy6YSVV0Cf), 2) kiyomi_fullmte-userdebug 8 fallbacks (http://shortn/_5UJCZ6PSjS)
Bug: b/371029823
Change-Id: I71651b00c27c0595a73ec1eb182277c55eaf929b
GitOrigin-RevId: 4f4b37ac14617145401b77709dda9738ab10aaf1