Avoid crash in reproxy when --version is specified as a flag
When a compile command for a C++ action is just /usr/bin/clang
--version, reproxy crashes because of the implicit dependency assumption
on there being a target file. This CL fixes the crash.

The linux kernel build calls clang to get its version to see if it would
be compatible with the build or not. If we want to use reclient there,
we would need this segfault fixed.

Bug: b/388825099
Tested: Ran with /usr/bin/clang --version
Change-Id: I03a26cd92a84218f40fa62025f7980e39f508cae
GitOrigin-RevId: 300a50978c3eba12a7c555ea3f31a84e9aafac8f