Cleaning up renovate configs
Make it so googleapis, and protobuf do not update on minor and patch releases.
Also make renovate ignore llvm_zlib and llvm_zstd packages as they are pinned to the version used by llvm.
Ignore abseil-cpp in renovate to keep it same version as goma fork.

We were already attempting to the things above but were missing a few rules, which hopefully this CL will amend.

Bug: b/400484161
Test: N/A
Change-Id: I054c23724968167928c42c3e5a3bc717d4720c14
GitOrigin-RevId: 85f4bf74c75cb63668b3737a09f8d330ceec8262